fixed gcc warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@2998 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtlib/core/Terrain.cpp
@@ -1315,7 +1315,7 @@ void vtTerrain::SetFogColor(const RGBf &color)
 
 void vtTerrain::SetFogDistance(float fMeters)
 {
-	m_Params.SetValueInt(STR_FOGDISTANCE, fMeters / 1000);
+	m_Params.SetValueInt(STR_FOGDISTANCE, (int) (fMeters / 1000));
 	if (m_bFog)
 		SetFog(true);
 }
